server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Cloud Orchestrator for UI Navigator - ADK powered"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "python-jose[cryptography]>=3.3.0",
    "pymongo[srv]>=4.10.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "sentry-sdk>=2.54.0",
    "ultralytics>=8.0.0",
    "easyocr>=1.7.0",
    "huggingface-hub>=0.20.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
